로그인, 회원가입페이지 공통 로직 모듈화
