+++ login/index.html
@@ -74,7 +74,7 @@
     </div>
   </div>
 
-<script src="index.js"></script>
+<script type="module" src="./index.js"></script>
 </body>
 
 </html>
