+++ login/style.css
@@ -163,9 +163,10 @@ input::placeholder {
 }
 
 .easyLoginText {
-font-size: var(--text-lg-16);
-font-weight: 500;
-line-height: 2.4rem; /*24px*/
+  font-size: var(--text-lg-16);
+  font-weight: 500;
+  line-height: 2.4rem; /*24px*/
+  color: var(--secondary-color-gray800);
 }
 
 .snsIc {
@@ -235,6 +236,8 @@ a.icImg {
 .modal {
   display: flex;
   position: relative;
+  width: 54rem;
+  height: 25rem;
   padding: 10.8rem 17.4rem 12.3rem 17.4rem ;
   background-color: var(--background-color-white);
   max-width: 54rem;
@@ -242,10 +245,10 @@ a.icImg {
 }
 
 .modalMessage {
-font-size: var(--text-lg-16);
-font-weight: 500;
-line-height: 1.909rem;
-color: var(--secondary-color-gray800);
+  font-size: var(--text-lg-16);
+  font-weight: 500;
+  line-height: 1.909rem;
+  color: var(--secondary-color-gray800);
 }
 
 .modalButton {
