+++ signup/index.html
@@ -37,7 +37,10 @@
 
       <div class="signupWrapper">
         <label for="nickname">닉네임</label>
-        <input id="nickname" name="nickname" placeholder="닉네임을 입력해주세요" autocomplete="on" required>
+        <div>
+          <input id="nickname" name="nickname" placeholder="닉네임을 입력해주세요" autocomplete="on" required>
+          <div id="emptyNicknameError" class="hide">닉네임을 입력해주세요.</div>
+        </div>
       </div>
 
       <div class="signupWrapper">
@@ -90,6 +93,6 @@
     </div>
   </div>
 
-  <script src="index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
