+++ signup/index.js
@@ -1,14 +1,8 @@
-const USER_DATA = [
-  { email: 'codeit1@codeit.com', password: "codeit101!" },
-  { email: 'codeit2@codeit.com', password: "codeit202!" },
-  { email: 'codeit3@codeit.com', password: "codeit303!" },
-  { email: 'codeit4@codeit.com', password: "codeit404!" },
-  { email: 'codeit5@codeit.com', password: "codeit505!" },
-  { email: 'codeit6@codeit.com', password: "codeit606!" },
-];
+import { USER_DATA, showErrorMessage, removeErrorMessage, changeFiledType } from '../module.js';
 
 // element setting
 const elInputEmail = document.querySelector('#email');
+const elInputNickname = document.querySelector('#nickname');
 const elInputPassword = document.querySelector('#password');
 const elInputConfirmPassword = document.querySelector('#confirmPassword');
 const elSignupBtn = document.querySelector('#signupButton');
@@ -16,26 +10,15 @@ const togglePasswordAll = document.querySelectorAll('.togglePassword');
 
 let emailComplete
 let passwordComplete
+let nicknameComplete
 let confirmPasswordComplete
 
-// 에러 메세지 표시
-function showErrorMessage(elInput, errorId) {
-  const visibleErrorMessage = document.querySelector(errorId)
-  visibleErrorMessage.style.display = 'block';
-  elInput.classList.add('borderRed');
-}
-// 에러 메세지 지우기
-function removeErrorMessage(elInput, errorId) {
-  const visibleErrorMessage = document.querySelector(errorId)
-  visibleErrorMessage.style.display = 'none';
-  elInput.classList.remove('borderRed');
-}
 // 이메일 입력상태 함수
 function validateEmail() {
   removeErrorMessage(elInputEmail, '#emptyEmailError');
   removeErrorMessage(elInputEmail, '#emailError');
   const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  emailComplete = false
+  emailComplete = false;
 
   if (elInputEmail.value.length === 0) {
     showErrorMessage(elInputEmail, '#emptyEmailError');
@@ -47,11 +30,23 @@ function validateEmail() {
   buttonActivation();
 }
 
+// 닉네임 입력상태 함수
+function validateNickname() {
+  removeErrorMessage(elInputNickname, '#emptyNicknameError');
+  nicknameComplete = false;
+
+  if (elInputNickname.value.length === 0) {
+    showErrorMessage(elInputNickname, '#emptyNicknameError');
+  } else {
+    nicknameComplete = true;
+  }
+}
+
 // 비밀번호 입력상태 함수
 function validatePassword() {
   removeErrorMessage(elInputPassword, '#emptyPasswordError');
   removeErrorMessage(elInputPassword, '#passwordError');
-  passwordComplete = false
+  passwordComplete = false;
 
   if (elInputPassword.value.length === 0) {
     showErrorMessage(elInputPassword, '#emptyPasswordError');
@@ -67,7 +62,7 @@ function validatePassword() {
 function validateConfirmPassword() {
   removeErrorMessage(elInputConfirmPassword, '#emptyConfirmPasswordError');
   removeErrorMessage(elInputConfirmPassword, '#confirmPasswordError');
-  confirmPasswordComplete = false
+  confirmPasswordComplete = false;
 
   if (elInputConfirmPassword.value.length === 0) {
     showErrorMessage(elInputConfirmPassword, '#emptyConfirmPasswordError');
@@ -81,7 +76,7 @@ function validateConfirmPassword() {
 
 // 버튼 활성화
 function buttonActivation() {
-  if(emailComplete && passwordComplete && confirmPasswordComplete) {
+  if(emailComplete && nicknameComplete && passwordComplete && confirmPasswordComplete) {
     elSignupBtn.disabled = false;
     elSignupBtn.addEventListener('click',checkUserData); // USER_DATA 이메일 유무 체크
   } else {
@@ -108,19 +103,11 @@ document.querySelector('.modalButton').addEventListener('click', () => {
 
 // togglePassword 클릭시 비밀번호 표시/숨기기
 togglePasswordAll.forEach(togglePassword => {
-  togglePassword.addEventListener('click', function() {
-    const passwordFiled = this.previousElementSibling;
-    if(passwordFiled.type === 'password') {
-      passwordFiled.type = 'text';
-      togglePassword.classList.add('togglePasswordHide');
-    } else {
-      passwordFiled.type = 'password';
-      togglePassword.classList.remove('togglePasswordHide');
-    }
-  })  
+  togglePassword.addEventListener('click', changeFiledType)  
 });
 
 /* event handling */ 
 elInputEmail.addEventListener('focusout',validateEmail);// 이메일 입력상태 호출
+elInputNickname.addEventListener('focusout',validateNickname);// 닉네임 입력상태 호출
 elInputPassword.addEventListener('focusout',validatePassword); // 비밀번호 입력상태 호출
 elInputConfirmPassword.addEventListener('focusout',validateConfirmPassword);// 비밀번호확인 입력상태 호출
